fastapi-cloud/requirements.txt
aiofiles==0.8.0
amqp==5.1.1
anyio==3.6.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
blis==0.7.7
cachetools==5.1.0
catalogue==2.0.7
celery==5.2.6
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cymem==2.0.6
Deprecated==1.2.13
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.3.0/en_core_web_lg-3.3.0-py3-none-any.whl
fastapi==0.78.0
gcloud==0.18.3
gensim==4.2.0
google-api-core==2.7.3
google-auth==2.6.6
google-cloud-core==2.3.0
google-cloud-storage==2.3.0
google-crc32c==1.3.0
google-resumable-media==2.3.2
googleapis-common-protos==1.56.1
greenlet==1.1.2
h11==0.13.0
httplib2==0.20.4
httptools==0.4.0
idna==3.3
Jinja2==3.1.2
joblib==1.1.0
kombu==5.2.4
langcodes==3.3.0
MarkupSafe==2.1.1
murmurhash==1.0.7
numpy==1.22.3
oauth2client==4.1.3
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
pathy==0.6.1
pdf2image==1.16.0
Pillow==9.1.1
preshed==3.0.6
prompt-toolkit==3.0.29
protobuf==4.21.0rc1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.8.2
PyMySQL==1.0.2
pyparsing==3.0.9
pytesseract==0.3.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-multipart==0.0.5
pytz==2022.1
PyYAML==6.0
redis==4.3.1
requests==2.27.1
rsa==4.8
scikit-learn==1.1.0
scipy==1.8.1
six==1.16.0
sklearn==0.0
smart-open==5.2.1
sniffio==1.2.0
spacy==3.3.0
spacy-legacy==3.0.9
spacy-loggers==1.0.2
SQLAlchemy==1.4.36
srsly==2.4.3
starlette==0.19.1
thinc==8.0.15
threadpoolctl==3.1.0
tqdm==4.64.0
typer==0.4.1
typing_extensions==4.2.0
urllib3==1.26.9
uvicorn==0.17.6
uvloop==0.16.0
vine==5.0.0
wasabi==0.9.1
watchgod==0.8.2
wcwidth==0.2.5
websockets==10.3
wrapt==1.14.1
